Remove unnecessary conversion

+++ internal/resources/snapshot_backup.go
@@ -317,7 +317,7 @@ func (s *SnapshotBackup) Update(ctx context.Context, req resource.UpdateRequest,
 		return
 	}
 
-	state.Retention = types.Int64Value(int64(refreshedState.Retention))
+	state.Retention = types.Int64Value(refreshedState.Retention)
 	state.Expiration = types.StringValue(refreshedState.Expiration)
 	state.RegionsToCopy = plan.RegionsToCopy
 	state.RestoreTimes = plan.RestoreTimes
